Move override_arch.py to _internal
