+++ certbot-apache/certbot_apache/entrypoint.py
@@ -6,7 +6,7 @@ from distutils.version import LooseVersion  # pylint: disable=no-name-in-module,
 from certbot import util
 
 from certbot_apache import configurator
-from certbot_apache import override_arch
+from certbot_apache._internal import override_arch
 from certbot_apache._internal import override_fedora
 from certbot_apache._internal import override_darwin
 from certbot_apache._internal import override_debian
